+++ models/spring-ai-vertex-ai-palm2/src/main/java/org/springframework/ai/vertexai/palm2/api/VertexAiPaLm2Api.java
@@ -364,7 +364,7 @@ public class VertexAiPaLm2Api {
 	 */
 	@JsonInclude(Include.NON_NULL)
 	public record Embedding(
-			@JsonProperty("value") List<Double> value) {
+			@JsonProperty("value") List<Float> value) {
 
 	}
